backend/requirement.txt
fastapi
sqlalchemy
pymysql
pydantic
uvicorn
python-dateutil
tensorflow==2.16.1